add libpisp
On RPi 5, (certain?) camera signals pass through PISP, which seems
to be a hardwared image processing block, which needs to be configured.
libpisp does it. Libpisp needs Niels Lohmanns JSON parser library.
In order for libcamera to work on BCM2712 devices, it needs to be
the patched version from the Raspberry github, and it also needs the
libpisp pipeline enabled

Signed-off-by: Thomas Sailer <tsailer@daedalean.ai>